refactor(workflow-tasks): move namespace component into library
